updated classes to restaurantRequest

+++ restaurant-backend/src/main/java/com/kristenwoodward/repository/RestaurantRepository.java
@@ -1,12 +1,9 @@
 package com.kristenwoodward.repository;
 
-import com.kristenwoodward.model.Restaurant;
+import com.kristenwoodward.model.persistence.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface RestaurantRepository extends JpaRepository<Restaurant, Integer> {
-
+public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
 }
